Refine the Dockerfile example
If programmers repeatedly rebuld the docker container of the operator
because they repeatedly make code changes in the operator code, then
"docker build" does not need to rebuild the unchanged base layers.
This changeset brings forward the installation of kopf, and therefore
the "pip install kopf" step can be omitted in this use-case.

On the other hand, if there's a new version of kopf, then (I think)
"docker build" won't recognize it should re-run pip and will continue
to use the cached layers.